+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - refactor(errors): improve pathFilter error message
 - fix(logger-plugin): fix missing target port
 - ci(package): npm package provenance
+- fix(logger-plugin): log target port when router option is used
 
 ## [v3.0.0](https://github.com/chimurai/http-proxy-middleware/releases/tag/v3.0.0)